requirements.txt
diffusers
transformers
pillow
torch
huggingface-hub
accelerate
requests
geopy
ee